Move example components from app to module
This is just easier for the code organization to keep it in one app.
