+++ laces/test/example/templates/components/hello-world.html
